Changing invocation to be in beforeEach/afterEach rather than testDone
