+++ testem.js
@@ -9,11 +9,14 @@ module.exports = {
     'Chrome'
   ],
   'browser_args': {
-      'Chrome': [
+    Chrome: {
+      mode: 'ci',
+      args: [
         '--disable-gpu',
         '--headless',
         '--remote-debugging-port=9222',
         '--window-size=1440,900'
-      ],
+      ]
+    },
   }
 };
