+++ tests/unit/sinon-sandbox-with-global-access-test.js
@@ -1,5 +1,6 @@
+/* global: sinon */
 import QUnit, { module, test } from 'qunit';
-import setupSinonSandbox, { createSandbox, restoreSandbox, setOptions } from 'ember-sinon-sandbox/test-support';
+import { createSandbox, restoreSandbox, setOptions } from 'ember-sinon-sandbox/test-support';
 
 module('Unit | ember-sinon-sandbox | With global access', {
   before() {
@@ -11,7 +12,19 @@ module('Unit | ember-sinon-sandbox | With global access', {
   }
 });
 
-test('stores sandbox created as module property', function(assert) {
+test('does not error when accessing the sinon global', function(assert) {
+  assert.expect(2);
+
+  createSandbox();
+
+  assert.ok(self.sinon, 'self.sinon is available');
+
+  assert.ok(self.sinon.spy(), 'sinon spy is available');
+
+  restoreSandbox();
+});
+
+test('sandbox created as module property and global sinon are the same sandbox instance', function(assert) {
   assert.expect(3);
 
   assert.notOk(QUnit.config.current.testEnvironment.sandbox);
@@ -38,80 +51,6 @@ test('stubbing out sandbox.create returns the already created sandbox', function
   restoreSandbox();
 });
 
-test('calling `sinon.sandbox.restore()` noops to ensure restoration is controlled', function(assert) {
-  assert.expect(2);
-
-  createSandbox();
-
-  this.sandbox.spy();
-
-  assert.equal(this.sandbox.fakes.length, 1);
-
-  this.sandbox.restore();
-
-  assert.equal(this.sandbox.fakes.length, 1);
-
-  restoreSandbox();
-});
-
-test('using useFakeTimers API continues to work', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  let clock = this.sandbox.useFakeTimers();
-
-  assert.ok(clock, 'The clock API continues to work after forced sandboxing.');
-
-  restoreSandbox();
-});
-
-test('using useFakeXMLHttpRequest API continues to work', function(assert) {
-  assert.expect(1);
-
-  let requests = [];
-
-  createSandbox();
-
-  let fakeXHR = this.sandbox.useFakeXMLHttpRequest();
-  fakeXHR.onCreate = (req) => {
-    requests.push(req);
-  };
-
-  let xhr = new XMLHttpRequest();
-  xhr.addEventListener("load", () => {});
-  xhr.open("GET", "http://www.example.org/example.txt");
-  xhr.send();
-
-  assert.equal(requests.length, 1, 'The fake XHR API continues to work after forced sandboxing.');
-
-  restoreSandbox();
-});
-
-test('using sinon.assert.* methods throws an error', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  assert.throws(() => {
-    this.sandbox.assert.calledOnce()
-  }, 'sinon.assert methods throw an error');
-
-  restoreSandbox();
-});
-
-test('using sinon.fakeServer.create throws an error', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  assert.throws(() => {
-    this.sandbox.fakeServer.create()
-  }, 'sandbox.fakeServer.create throws and error');
-
-  restoreSandbox();
-});
-
 test('ensures sandbox is restored correctly', function(assert) {
   assert.expect(1);
 
@@ -132,32 +71,4 @@ test('ensures sandbox instances are different for each test', function(assert) {
     previousSandbox = Object.assign({}, self.sinon);
     restoreSandbox();
   }
-});
-
-test('configuring setup/restore', function(assert) {
-  assert.expect(4);
-
-  let testStartCalled = false;
-  let testDoneCalled = false;
-
-
-  let options = {
-    QUnit: {
-      testStart(callback) {
-        testStartCalled = true;
-        assert.equal(callback, createSandbox);
-      },
-
-      testDone(callback) {
-        testDoneCalled = true;
-        assert.equal(callback, restoreSandbox);
-      }
-    },
-    errorOnGlobalSinonAccess: false
-  };
-
-  setupSinonSandbox(options);
-
-  assert.ok(testStartCalled, 'testEnvironment.testStart is called');
-  assert.ok(testDoneCalled, 'testEnvironment.testDone is called');
 });
